Improved capture panel

+++ digiCamControl.LightBox.Plugins/CapturePanelViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
@@ -15,8 +16,14 @@ namespace digiCamControl.LightBox.Plugins
     {
         public RelayCommand CaptureCommand { get; set; }
         public Session Session => ServiceProvider.Instance.Session;
-        private List<PhotoCapturedEventArgs> _fileStack = new List<PhotoCapturedEventArgs>();
-        private bool _captureDone = false;
+        public ICameraDevice CameraDevice => ServiceProvider.Instance.DeviceManager.SelectedCameraDevice;
+        private ConcurrentQueue<PhotoCapturedEventArgs> _fileStack = new ConcurrentQueue<PhotoCapturedEventArgs>();
+        private bool _transferInProgress;
+        private int _totalFiles;
+        private int _filesProgress;
+        private int _captureProgress;
+        private bool _captureInProgress;
+        private double _waitProgress;
 
         public bool CaptureWithNoAf
         {
@@ -48,6 +55,32 @@ namespace digiCamControl.LightBox.Plugins
             }
         }
 
+        public double WaitProgress
+        {
+            get { return _waitProgress; }
+            set
+            {
+                _waitProgress = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        /// <summary>
+        /// Indicate if a capture process (whole series is started)
+        /// </summary>
+        public bool CaptureInProgress
+        {
+            get { return _captureInProgress; }
+            set
+            {
+                _captureInProgress = value;
+                RaisePropertyChanged();
+                RaisePropertyChanged(() => IsFreeToCapture);
+            }
+        }
+
+        public bool IsFreeToCapture => !CaptureInProgress;
+
         public bool CaptureTransferAfterCapture
         {
             get { return Session.Variables.GetBool("CaptureTransferAfterCapture"); }
@@ -58,20 +91,75 @@ namespace digiCamControl.LightBox.Plugins
             }
         }
 
+        public bool LeaveFileAfterTransfer
+        {
+            get { return Session.Variables.GetBool("LeaveFileAfterTransfer"); }
+            set
+            {
+                Session.Variables["LeaveFileAfterTransfer"] = value;
+                RaisePropertyChanged();
+            }
+        }
 
         public CapturePanelViewModel()
         {
             CaptureCommand = new RelayCommand(Capture);
             if (!IsInDesignMode)
             {
+                
                 ServiceProvider.Instance.DeviceManager.PhotoCaptured += DeviceManager_PhotoCaptured;
                 if (CaptureCount < 1)
                     CaptureCount = 1;
                 if (Math.Abs(CaptureWait) < 0.01)
                     CaptureWait = 1;
             }
+            else
+            {
+                ServiceProvider.Instance.Session = new Session();
+            }
 
         }
+
+        public bool TransferInProgress
+        {
+            get { return _transferInProgress; }
+            set
+            {
+                _transferInProgress = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        public int TotalFiles
+        {
+            get { return _totalFiles; }
+            set
+            {
+                _totalFiles = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        public int FilesProgress
+        {
+            get { return _filesProgress; }
+            set
+            {
+                _filesProgress = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        public int CaptureProgress
+        {
+            get { return _captureProgress; }
+            set
+            {
+                _captureProgress = value;
+                RaisePropertyChanged();
+            }
+        }
+
 
         private void Capture()
         {
@@ -84,54 +172,70 @@ namespace digiCamControl.LightBox.Plugins
             try
             {
                 TransferFiles();
+                CaptureInProgress = true;
                 if (CaptureTransferAfterCapture)
                     ServiceProvider.Instance.DeviceManager.SelectedCameraDevice.CaptureInSdRam = false;
 
                 ServiceProvider.Instance.OnMessage(Messages.StopLiveView);
                 for (int i = 0; i < CaptureCount; i++)
                 {
-                    _captureDone = false;
+                    CaptureProgress =  i+1;
                     if (CaptureWithNoAf)
-                        ServiceProvider.Instance.DeviceManager.SelectedCameraDevice.CapturePhotoNoAf();
+                        Utils.ExecuteWithRetry(
+                            ServiceProvider.Instance.DeviceManager.SelectedCameraDevice.CapturePhotoNoAf);
                     else
-                        ServiceProvider.Instance.DeviceManager.SelectedCameraDevice.CapturePhoto();
-                    if (i < CaptureCount - 1)
-                        Thread.Sleep((int) (CaptureWait * 1000));
-                    while (!_captureDone)
+                        Utils.ExecuteWithRetry(ServiceProvider.Instance.DeviceManager.SelectedCameraDevice.CapturePhoto);
+                    if (i < CaptureCount - 1 && CaptureWait > 0)
                     {
-                        Thread.Sleep(1);
+                        WaitProgress = 0;
+                        while (WaitProgress < CaptureWait-0.05)
+                        {
+                            WaitProgress += 0.050;
+                            Thread.Sleep((int) (Math.Min(0.050, CaptureWait - WaitProgress) * 1000));
+                        }
+                        WaitProgress = CaptureWait;
                     }
                 }
-
-                TransferFiles();
+            }
+            catch (DeviceException  deviceException)
+            {
+                ServiceProvider.Instance.OnMessage(Messages.Message, deviceException.Message);
+                Log.Error("Capture error", deviceException);
             }
             catch (Exception e)
             {
                 Log.Error("Capture error", e);
             }
+            CaptureProgress = 0;
+            CaptureInProgress = false;
+            TransferFiles();
         }
 
         private void DeviceManager_PhotoCaptured(object sender, PhotoCapturedEventArgs eventargs)
         {
-            _fileStack.Add(eventargs);
+            _fileStack.Enqueue(eventargs);
             if (!CaptureTransferAfterCapture)
                 TransferFiles();
-            _captureDone = true;
         }
 
 
 
         private void TransferFiles()
         {
-            foreach (var eventargs in _fileStack)
+            TransferInProgress = true;
+            PhotoCapturedEventArgs eventargs;
+            TotalFiles = _fileStack.Count;
+            while (_fileStack.TryDequeue(out eventargs))
             {
                 try
                 {
+                    FilesProgress++;
                     string fileName = Path.Combine(Settings.Instance.TempFolder,
                         Path.GetRandomFileName() + Path.GetExtension(eventargs.FileName));
                     Utils.CreateFolder(fileName);
                     eventargs.CameraDevice.TransferFile(eventargs.Handle, fileName);
-                    eventargs.CameraDevice.DeleteObject(new DeviceObject() { Handle = eventargs.Handle });
+                    if (!LeaveFileAfterTransfer)
+                        eventargs.CameraDevice.DeleteObject(new DeviceObject() {Handle = eventargs.Handle});
                     eventargs.CameraDevice.ReleaseResurce(eventargs.Handle);
                     ServiceProvider.Instance.OnMessage(Messages.ImageCaptured, fileName);
                 }
@@ -140,7 +244,7 @@ namespace digiCamControl.LightBox.Plugins
                     Log.Debug("File transffer error", e);
                 }
             }
-            _fileStack.Clear();
+            TransferInProgress = false;
         }
 
     }
